+++ packages/core/src/core/content/icons/List.tsx
@@ -1,0 +1,45 @@
+import { memo } from 'react'
+import { Line as _Line, Rect as _Rect, Svg as _Svg } from 'react-native-svg'
+import { themed } from './themed'
+
+const Icon = (props) => {
+  const { color = 'black', size = 24, ...otherProps } = props
+  return (
+    <_Svg viewBox="0 0 256 256" {...otherProps} height={size} width={size}>
+      <_Rect width="256" height="256" fill="none" />
+      <_Line
+        x1="40"
+        y1="128"
+        x2="216"
+        y2="128"
+        stroke={`${color}`}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="16"
+      />
+      <_Line
+        x1="40"
+        y1="64"
+        x2="216"
+        y2="64"
+        stroke={`${color}`}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="16"
+      />
+      <_Line
+        x1="40"
+        y1="192"
+        x2="216"
+        y2="192"
+        stroke={`${color}`}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="16"
+      />
+    </_Svg>
+  )
+}
+Icon.displayName = 'ListRegular'
+const ListRegular = memo(themed(Icon))
+export { ListRegular }