+++ packages/core/src/core/content/icons/TextUnderline.tsx
@@ -1,0 +1,34 @@
+import { memo } from 'react'
+import { Line as _Line, Path as _Path, Rect as _Rect, Svg as _Svg } from 'react-native-svg'
+import { themed } from './themed'
+
+const Icon = (props) => {
+  const { color = 'black', size = 24, ...otherProps } = props
+  return (
+    <_Svg viewBox="0 0 256 256" {...otherProps} height={size} width={size}>
+      <_Rect width="256" height="256" fill="none" />
+      <_Line
+        x1="40"
+        y1="216"
+        x2="216"
+        y2="216"
+        fill="none"
+        stroke={`${color}`}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="16"
+      />
+      <_Path
+        d="M64,56v64a64,64,0,0,0,128,0V56"
+        fill="none"
+        stroke={`${color}`}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="16"
+      />
+    </_Svg>
+  )
+}
+Icon.displayName = 'TextUnderlineRegular'
+const TextUnderlineRegular = memo(themed(Icon))
+export { TextUnderlineRegular }