+++ packages/rich-text/src/LmTipTapMenuBar.tsx
@@ -1,14 +1,14 @@
 import { Editor } from '@tiptap/react'
 import {
-  ArrowClockwise,
-  ArrowCounterClockwise,
-  List,
-  ListNumbers,
-  Quotes,
-  TextBolder,
-  TextItalic,
-  TextStrikethrough,
-} from 'tamagui-phosphor-icons'
+  ArrowClockwiseRegular,
+  ArrowCounterClockwiseRegular,
+  ListNumbersRegular,
+  ListRegular,
+  QuotesRegular,
+  TextBolderRegular,
+  TextItalicRegular,
+  TextStrikethroughRegular,
+} from '@tamagui-extras/core'
 import { Button, ColorTokens, XStack } from 'tamagui'
 import { LmSelect } from '@tamagui-extras/form'
 
@@ -25,20 +25,24 @@ export function LmTipTapToolbar({ editor }: { editor: Editor | null }) {
         chromeless
         onPress={() => editor.chain().focus().toggleBold().run()}
         disabled={!editor.can().chain().focus().toggleBold().run()}
-        icon={<TextBolder color={editor.isActive('bold') ? activeThemeColor : undefined} />}
+        icon={<TextBolderRegular color={editor.isActive('bold') ? activeThemeColor : undefined} />}
       />
       <Button
         chromeless
         onPress={() => editor.chain().focus().toggleItalic().run()}
         disabled={!editor.can().chain().focus().toggleItalic().run()}
-        icon={<TextItalic color={editor.isActive('italic') ? activeThemeColor : undefined} />}
+        icon={
+          <TextItalicRegular color={editor.isActive('italic') ? activeThemeColor : undefined} />
+        }
       />
       <Button
         chromeless
         onPress={() => editor.chain().focus().toggleStrike().run()}
         disabled={!editor.can().chain().focus().toggleStrike().run()}
         icon={
-          <TextStrikethrough color={editor.isActive('strike') ? activeThemeColor : undefined} />
+          <TextStrikethroughRegular
+            color={editor.isActive('strike') ? activeThemeColor : undefined}
+          />
         }
       />
 
@@ -69,33 +73,39 @@ export function LmTipTapToolbar({ editor }: { editor: Editor | null }) {
       />
       <Button
         chromeless
-        icon={<List color={editor.isActive('bulletList') ? activeThemeColor : undefined} />}
+        icon={<ListRegular color={editor.isActive('bulletList') ? activeThemeColor : undefined} />}
         onPress={() => editor?.chain().focus().toggleBulletList().run()}
         disabled={!editor?.can().chain().focus().toggleBulletList().run()}
       />
       <Button
         chromeless
-        icon={<ListNumbers color={editor.isActive('orderedList') ? activeThemeColor : undefined} />}
+        icon={
+          <ListNumbersRegular
+            color={editor.isActive('orderedList') ? activeThemeColor : undefined}
+          />
+        }
         onPress={() => editor?.chain().focus().toggleOrderedList().run()}
         disabled={!editor?.can().chain().focus().toggleOrderedList().run()}
       />
 
       <Button
         chromeless
-        icon={<Quotes color={editor.isActive('blockquote') ? activeThemeColor : undefined} />}
+        icon={
+          <QuotesRegular color={editor.isActive('blockquote') ? activeThemeColor : undefined} />
+        }
         onPress={() => editor.chain().focus().toggleBlockquote().run()}
         disabled={!editor.can().chain().focus().toggleBlockquote().run()}
       />
 
       <Button
         chromeless
-        icon={<ArrowCounterClockwise />}
+        icon={<ArrowCounterClockwiseRegular />}
         onPress={() => editor.chain().focus().undo().run()}
         disabled={!editor.can().chain().focus().undo().run()}
       />
       <Button
         chromeless
-        icon={<ArrowClockwise />}
+        icon={<ArrowClockwiseRegular />}
         onPress={() => editor.chain().focus().redo().run()}
         disabled={!editor.can().chain().focus().redo().run()}
       />