+++ packages/video/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/video
 
+## 1.2.0
+
+### Minor Changes
+
+- Remove dependencies of tamagui-phosphor-icons in all packages
+
 ## 1.1.3
 
 ### Patch Changes