+++ packages/youtube/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tamagui-extras/youtube
 
+## 1.1.0
+
+### Minor Changes
+
+- Remove dependencies of tamagui-phosphor-icons in all packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/core@1.1.0
+
 ## 1.0.7
 
 ### Patch Changes